Fix allowing moment_kinetics.jl, post_processing.jl to be run directly
With the recent update they could be run directly, but errored when
running from the 'driver' file because the function call could not be
compiled correctly when importing the files. Fix is to surround the
function call with a check that the file is being run directly as the
'main' program, so it is not called when importing.