Fix long tests in CI
Skip a workaround used by the 'julia-runtest' action for the case when
the Julia 'General Registry' is not perfectly up to date - requires a
file from that action, so rather than copying the file just skip as it
doesn't seem to be too important.

Remove `force_latest_compatible_version` argument which caused an error.