Specialised versions of integrate to avoid temporary arrays
Most of the times we want to integrate something, we are taking a
velocity moment of the form `f * v^n * weight`. New versions allow f and
v weights to be passed as separate arguments to avoid having to
calculate `f * v^n` in a temporary array when passing the arguments in.