Redesign regression tests in sound_wave_test.jl
The fitted growth rate and frequency are too sensitive to rounding
errors, even with relative tolerance set at 2.0e-8, keep needing to
update the output of one test or another. This commit changes the
regression test to instead check a subset of the final values of phi -
these seem to be more consistent, which should make regression testing
more robust.