Add debugging features
Provides a set of macros that can be used to wrap blocks of code used
only for debugging. The macro blocks are activated if _debug_level is
set to a high enough level. Note _debug_level can only be changed before
precompiling/compiling the code, so even if using Revise.jl, need to
restart the REPL for changes to take effect.