Add function to plot strong-scaling performance in performance-tests/
Plots both run time and efficiency=(ideal run time)/(run time).

Includes estimate of efficiency with 'actual' load balance when plotting
performance, to check communication efficiency: calculate a 'less ideal'
scaling, that uses the 'actual' load balance (i.e. largest number of
grid points on a single process) to calculate the expected run time.

The 'actual load balance' is an estimate. At the moment it uses the
number of points in the z-direction for the 'grid size' which is split
up, which is true for probably more than half of the loops in the code,
but not the ones in z_advection!(). Load balance should be better when
the vpa-direction is split (as there are more points in this dimension
in the sound_wave.jl test case), so this is a pessimistic estimate of
load balance, so an optimistic estimate of communication efficiency.

Also fix error bars in plot_performance.jl: Need both entries in yerror
argument tuple to be vectors, otherwise only the first value in the
single vector is used.

Note: If the range does not include at least 2 major ticks, the axis
labels get messed up. This commit is a hacky workaround used for the
history plots. Not applied to strong scaling plots (even though the
issue occasionally affects them) because too many of them look awful
with it.