Use more options in precompile_run.jl
Should mean more of the code gets compiled into the `.so`.

When precompiling dependencies, just use a short run with one set of
parameters because that should be sufficient to make all the
dependencies compile.

Also put the precompilation run output in a temporary directory so we do
not have to worry about deleting it, or conflicting with real output.