Add timeouts to CI tests, use `--debug 1` and try to speed up parallel tests
Add timeouts so CI jobs do not hang for too long. This should make it
quicker to see output if jobs are failing (some output sometimes gets
hidden while jobs are still running).

When running in parallel, all processes compile code, which wastes
resources. Therefore precompile (in serial) for the parallel test jobs.
Also means that precompilation gets checked somewhere, which is good.

Use `--check-bounds=no` for parallel tests.

Also use debug level 1 for parallel jobs, because it means that we exit
cleanly on all processes if there is an error on any process, so should
help avoid runs hanging for a long time.