Fix performace of parallel tests in CI
Use openmpi and pass options to allow oversubscription and prevent
terrible performace when oversubscribing.

Supposedly there is a similar option to prevent terrible performance
when oversubscribing for mpich (which MPI.jl uses by default) but it was
not working. Not sure if this was a mistake by me, or the option needs
to be given when configuring/compiling mpich, or something else...