Force N_e to be positive in update_phi!()
Physically, N_e must be positive, because jpar_i into the lower boundary
should always be negative (flow in the -'ve z-direction). However,
during initialization it may be that jpar_i=0, or in an underresolved
simulation it may that jpar_i>=0. Therefore, force N_e to be always
positive (>=1.0e-16) to avoid ever taking log() of a non-positive
number.