Better fit error calculation for phi-fitting routines
Re-define the expression used to normalise the relative error to be
`(abs(function)+abs(fit))^2` so that it is always greater than or equal
to the error `(function-fit)^2`, even when `function` and `fit` have
opposite signs.

Use a 3-point moving average of the normalisation factor to avoid giving
too much weight to errors on points that happen to be close to a
zero-crossing (where the normalisation factor is close to zero).