Simplify loop macros - loop_ranges[] depends on region
loop_ranges[] contains ranges for loops over each of the dimensions
(LoopRanges is a simpler struct than before). The LoopRanges instance in
loop_ranges[] is swapped out for the appropriate new one in the
begin_*_region() functions. That means the loop macros no longer need to
know what kind of region they are in.