Use begin_s_region() for moment updates
Because the moment updates in `continuity_equation!()`, etc., need to
take z-derivatives, they can only parallelize over species. Previously
this was done by using `@s_z_loop_s`, and then
`if 1 in loop_ranges[].s_z_loop_z`, in order to avoid changing the loop
type. It seems clearer to instead completely separate the part of
euler_time_advance!() that updates the distribution function (by moving
the boundary condition application from the bottom of
euler_time_advance!() to the end of the distribution function update),
and then use `begin_s_region(no_synchronize=true)` before the moment
update, which then uses `@s_loop`. It is possible to use the
`no_synchronize=true` argument to skip the _block_synchronize() call at
that point because before it only the distribution function is updated
and after it only the moments are updated, so there is no possibility of
a race condition.