Add debugging function debug_check_shared_memory()
Refactor debug_check_shared_array() out as a separate function, and use
it in debug_check_shared_memory().

debug_check_shared_memory() is only available when a high enough
debugging level is set (so @debug_shared_array is active), and is not
used normally but can be added to pin down where exactly a shared-memory
error occurred.