Move _block_synchronize() call outside r-loop in update_phi!()
Should not have `block_synchronize()` calls inside a parallelized loop -
it could end up being called different numbers of times on different
processes, which would be an error.

This change (which also simplifies the structure of the code a bit) was
pretty much made in `master` but reverted (by mistake) in merging.