Activate split_2_moments nonlinear sound wave tests
Previously these test failed because the charge exchange collisions did
not account for shifted grids when upar is evolved. Now the
split_2_moments tests can pass.

To get reasonably consistent results, it is necessary to increase the
resolution of the Chebyshev tests (and decrease the timestep).
Therefore, new expected results have been calculated/saved.

The output of tests with evolve_moments_parallel_flow==true requires
interpolation from the wpar grid to a upar grid (where the expected
results are saved). The order of the z and vpa interpolations of the
distribution function has been swapped because it seems more natural to
interpolate first in vpa using upar from the simulation (with no
interpolation), and then interpolate in z. Previously interpolation was
done first in z and then in vpa.

The split_3_moments tests are still expected to fail because the charge
exchange operator does not yet account for re-scaling of the wpar grid
by a z-dependent vth.