Fix bug that was skipping the tests in test/harrisonthompson.jl
`global_rank` is a `Ref` variable, so the check should be
`global_rank[] == 0` (testing the actual value). Previously had
`global_rank == 0`, which always evaluates to `false` so that the actual
tests were skipped.

Also update the tolerance for the regression test, because
`rtol = 1.e-14` was apparently too tight to allow for differences in
rounding errors, etc.