Make begin_*_region() return immediately if code already in that region
Intention is to allow begin_*_region() calls to be placed logically
(e.g. at the beginning of functions) without having to eliminate
redundant calls. This should allow the code to be written so that is not
necessary to search back through what was called previously to find the
most recent begin_*_region() call in order to know what the current
region is.