Put begin_*_region() functions inside each 'term'
Move `begin_*_region()` calls down inside the functions where they apply,
e.g. `begin_s_r_z_region()` is called at the beginning of
vpa_advection!() to declare that the body of that function is in a
`s_r_z` parallel region. As `begin_*_region()` is now zero cost if the
region is not changed, redundant calls (when they happen) are fine.