MMS test: use "zero" bc for vpa, non-zero upar, correct f normalization
Tricky to write down manufactured solution that exactly obeys "periodic"
bc in vpa, so switch to "zero" bc instead. Add prefactor to manufactured
f so it exactly obeys this bc at +/- Lvpa/2.

Add odd component to manufactured f so upar will be non-zero (and
non-constant). Done in a slightly strange way to ensure the additional
component is odd, but never causes f to be negative anywhere. Tried
adding non-zero upar by using a shifted Maxwellian type function, but
this introduces erf() functions with non-constant arguments to the
normalization correction (see below) which can be very expensive to
evaluate, making the code slow.

Ensure that the manufactured solution when integrated on
-Lvpa/2<vpa<Lvpa/2 and 0<vperp<Lvperp is exactly equal to the
manufactured n. This is necessary to ensure convergence, as the code
only evaluates the numerical integral within these limits.