Update expected values for wall bc test
The results have changed only because no boundary condition in v-space
is now applied to the neutrals, where previously periodic boundary
conditions were applied as for the charged particles. A boundary
condition is not needed for the neutrals as they have no advection in
v-space (i.e. no acceleration).

The changes are anyway small <1e-7. If we had set up the test with
`vpa.L = 16` instead of `vpa.L = 8`, then the differences would have
been <1.1e-12.