Fix cartesian_dfni_sym(), gyroaveraged_dfnn_sym() to include norm
Construct vrvzvzeta_dfni by substituting for vperp and vpa in terms of
vzeta, vr, and vz. This means any normalisation, etc. that is applied to
dfni is automatically carried over to vrvzvzeta_dfni.

Similarly construct gav_dfnn by substituting for vzeta, vr, and vz in
terms of vperp and vpa. Requires using a trick vzeta=vperp=-vr which
effectively gyroaverages by getting rid of the particular form of odd
components in vzeta and vr that were defined in dfnn_sym.