Handle ion wall bc with evolving upar in enforce_moment_constraints!()
When enforcing moment constraints the old distribution function (which
is assumed to already satisfy the constraints) is used to update the new
distribution function. Before this update, this could lead to the new
distribution function not quite obeying the boundary condition after
enforcing the constraints when upar is evolved, if upar had changed in
the wrong direction. Fix this problem by first applying the ion sheath
boundary condition with the new upar to the old distribution function,
then forcing it to obey the moment constraints.