Make update in neutral wall bc consistent with integration routines
integrate_over_positive_vpa() and integrate_over_negative_vpa() treat a
point where 'abs(vpa) < abs(zero)' specially (where zero is a small
constant to deal with finite precision): integration weights are
modified so such a point contributes half to the integral over positive
vpa and half to the integral over negative vpa. So if there is such a
point, the neutral pdf there needs to be set to half the incoming pdf
plus half the Knudsen cosine distribution, so that the integral over the
result times vpa vanishes exactly.