Use tighter and consistent tolerances to allow for finite precision
In a couple of places a constant 'zero' is defined, to check whether
something is equivalent to 0, but with finite precision. The same
constant shoud be used everywhere for consistency (should probably make
this a global, maybe in type_definitions.jl at some point...), and
because we want to be able to converge to very small errors (<1e-12 or
so at least) it makes sense to have 'zero' be smaller than it was
previously defined.