This commit reintroduces the Dirichlet boundary condition for the radial dimension. The boundary condition is fixed in time and imposed every time step. The boundary condition is imposed through distribution function arrays for the values of the neutral and charged particle and normalised distribution functions on r.n = 1 and r.n = end. These arrays are initialised after the distribution functions themselves are initialised, using the values of the initial condition on ir = 1 and ir = end. This means that the initial condition will always satisfy the boundary condition by construction.
Although the standard initialisation options do not allow for radial variation in the initial condition, we can test this implementation using the method of manufactured solutions framework. The modifications to manufactured_solutions.jl add a Dirichlet boundary condition test for both the z_bc = "periodic" and z_bc = "wall" test options. Initial inspection of the results suggest that the implementation is operating correctly.

The z_bc = "periodic" option tests the neutral and ion boundary conditions by having a linear spatial function in r for neutrals and a linear spatial function in r times a sine in z for ions. The sine in z is needed to generate a radial electric field (causing r advection).

The z_bc = "wall" option uses the previously constructed test that was compatible with r_bc = "periodic". The only modification is to replace nplus, nminus and nzero with functions that vary linearly in r. This test may be used to isolate the outstanding issue with the 2D-3V wall test with z_bc = "wall". Here we can force the radial electric field to be a constant in r, unlike in the r_bc = "periodic" case, where it was forced to be zero, or vary with r.