Make neutral boundary condition compatible with moment constraints
Previously, enforcing moment constraints would modify the neutral
distribution function at the sheath boundaries so that it no longer
obeys the boundary condition (in the sense that applying the boundary
condition again would modify the distribution function).

This is not an issue for the ions, because their boundary condition just
sets values to 0, and enforce_moment_constraints!() (with a bit of
special handling for boundary points on the input distribution function)
only adds contributions proportional to a distribution function that is
also 0 where the bc requires it to be 0.

Fix this for neutrals by applying the boundary condition in such a way
that the moment constraints are exactly satisfied, so that calling
enforce_moment_constraints!() after applying the bc would do nothing.