Introduce damping in last elements before vpa boundaries
Aim to suppress numerical instabilities that show up as oscillations
there.

Damping rate is adjusted according to dt.  Probably not a good idea in
the long run, but when trying out different dt's, it doesn't make sense
to have a damping rate faster than 1/dt, so set relative to 1/dt for
now.

Disabled for the moment.