One sided stencils at zero/both_zero/wall bndries for finite difference
At boundaries where the boundary condition sets a value, it should be OK
to use one-sided differences at the boundary point to calculate the
derivative (or unbalanced derivatives at the point next to the boundary
for higher order methods where this is necessary). This is more
consistent with the Chebyshev spectral element way of dealing with the
boundaries, and should be more numerically stable as artificially
setting values for ghost points can impose a discontinuity in the
derivative between the boundary point and interior points.