Allow passing file directly to analyze_and_plot_data()
Previously, could only pass an output directory to
analyze_and_plot_data(). Now that it is possible to restart, it is also
useful to be able to point directly to a specific output file. So now,
if the argument to analyze_and_plot_data() is a directory, the old
behaviour applies and it will analyze the default output file, but if a
file is passed, then that file will be analyzed.