Optionally make some plots at each output step to monitor progress
Can be useful to monitor an ongoing simulation. This cannot be done by
reading the output file, because opening the NetCDF file from two
programs at once may corrupt it and crash either/both programs. Instead,
if `runtime_plots = true` is passed in the input, a .png file with a
bunch of plots of moments and distribution functions is produced at each
output step. The plots are created in serial, and may take some time
(~10s?) for high-resolution simulations.