Nicer initial conditions for wall-bc cases
Create initial conditions by linearly combining the initial pdfs from
either wall. Avoids having a blob of almost-stationary particles in the
middle of the domain at t=0, which take a long time to advect away and
cause numerical problems due to steep gradients as the pdf that is
initeially roughly circular in z-vpa 'shears' due to parallel advection.