Move redundant _block_synchronize() checks into separate subdirectory
The checks for redundant _block_synchronize() calls require all
dimensions to be split over processes, so they require a large number of
processes to run in 2D2V/2D3V mode. Using that many processes, even if
oversubscribing the physical cpu cores, requires a lot of memory, so is
not feasible on CI servers or regular desktop/laptop computers.
Therefore the tests intended for checking for redundant
_block_synchronize() calls have been separated into a separate
subdiretory, to allow the standard debug checks to be modified so that
they can run on only 2 processes.

Also disable the redundant _block_synchronize() checks in CI.