Use system Python for PyCall on all platforms in CI jobs
Using the system Python is the default on Linux, but not macOS (or
Windows). This commit explicitly sets the ENV["PYTHON"] variable so that
all platforms use the system Python. This was needed to get PyCall to
work correctly on macOS - previously Julia's internal Conda-installed
Python had linking errors when using matplotlib.