Fix figure sizes
When creating figures for multiple runs, we need to set the figure size
explicitly for 2d plots (it should get wider when there are more runs to
avoid squashing the figures). This commit changes the settings so that
the size of the plot for each run matches the Plots.jl default size for
a single plot.