Ensure up-to-date moments are used everywhere
density, upar and ppar are updated in the scratch structs, while qpar
and vth are updated in the moments struct. Ensure density, upar and ppar
are always used from `fvec_in`, not from `momenst`. [density, upar and
ppar are only updated in the moments struct at the end of each RK
timestep, not for each RK stage.]