Example input files for wall-bc test case
Modifies parameters in wall-bc_cheb.toml, adds moment-kinetic versions
with 1, 2 and 3 moments split.

All cases run successfully and give 'the same' results when restarted
for 10 runs, with vpa_dissipation set to 1.0e-1 on the first run, 1.0e-2
on the second, 1.0e-3 on the third and 0.0 for the remaining restarts.
'Full-f', split-1 (evolving density) and split-2 (evolving density and
upar) give identical (by eye at least) results. Split-3 (evolving
density, upar and ppar) is not identical as numerical dissipation
(diffusion in vpa) in this case cannot heat the ions and neutrals
(because the dissipation as implemented does not modify the 'energy
equation'); however the final state after the 10th run seems to be
converging to the same result as the other runs, apart possibly from
very close to the wall where there is some numerical instability (?) at
large incoming vpa, which seems to have a small effect on the results.