Fix error handling for serial case
When running in serial, it is possible to clean up file I/O after
catching an error (as we do not have to call MPI.Abort() to get all
processes to stop). Previously this failed as mk_state was defined
within the try... block, so was undefined in the catch.... Now define
`mk_state = nothing` before the try...catch... to fix the bug.