Overlap communications in reconcile_element_boundaries_MPI!()
'send_buffer' was not strictly necessary as the endpoints are already
passed to the function in contiguous buffers, so repurpose (and rename)
'send_buffer' as 'receive_buffer1' (and
'receive_buffer'->'receive_buffer2') and start both 'sending up' and
'sending down' communications at the same time, with one MPI.Waitall()
call finalising both. This should be a (slight?) optimization.