MH's fix for numerical dissipation in source terms. To squash with rest of 'new MMS source for wall-bc' commits?
Refactor how numerical dissipation terms appear in the manufactured sources for MMS tests. This commit makes sure that diffusion terms are only included when the sign of the diffusion coefficient is positive, in line with the usage of the numerical dissipation in the main code. A Bool flag is added for possible later use switching on or off the numerical dissipation in the source. It is plausible that the manufactured source should not include the numerical dissipation if numerical dissipation is effective at smoothing the desired solution with a vanishingly small viscosity.