Fix error handling when running on multiple processes
MPI.free() is a collective call, and previously there was a
try...finally... that tried to call cleanup_moment_kinetics!() (which
calls MPI.free()) on any error. When running on multiple processes,
this meant the code would hang if an error occured on some processes and
not on others. Removing the try...finally... allows errors to trigger
MPI.Abort() as intended.