Fix merge of advance.*_diffusion flags, 'zero' boundary conditions
As implemented in tho 2D2V branch, the 'zero' boundary condition should
apply a Dirichlet boundary condition only at the upwind boundary when
there is advection but no diffusion, or at both boundaries when there is
diffusion. This commit enforces this behaviour in the merged branch, and
applies it consistently to all r, vpa, and vz diffusion. The z direction
is not included as if there is a wall boundary the outflow is very
strong, and we definitely do not want to impose a Dirichlet boundary
condition there.