Update filename for distributed, multiple-file mode
Swap order of the ".dfns" or ".moments" suffix and the ".<iblock>" part
(which is present when multiple files are being written, but not when
parallel I/O is being used). This makes it easier to extract the
'iblock' component from the filename, as we know that it is always
preceded by either ".dfns" or ".moments", whereas the user defined
prefix might end in something like ".2" by chance, which could be
mistaken for an 'iblock' suffix in the naming scheme before this update.