Move boundary conditions to rk_update!()
Ensures that the updated distribution function produced by the RK stage
has boundary conditions consistently applied. The previous version had
an updated distribution function calculated as the sum of 3 different
distribution functions each of which had the boundary condition applied,
but as these boundary conditions were applied with slightly different
`phi`s, this is not the same thing.