Remove the `run_type` global variable
`run_type` was originally used as a setting to turn timer output on/off,
etc. However, having a const global variable means that moment_kinetics
has to be re-(pre-)compiled if the value is changed, which is
inconvenient.

Refactor so that timing info can be printed if a TimerOutput object is
passed to run_moment_kinetics(), so `run_type` is not needed for this.

Parameter scans will be re-implemented without using `run_type`.