Truncate names when making run names for parameter scan
HDF5 seems to have a buffer overrun if file names are too long. Truncate
the scanned variable names to 3 characters to make this less likely to
happen.

Since the parameter scan saves its runs under a unique subdirectory
(named after the scan input file), there is no chance of accidentally
conflicting with the name of a run from a different scan.