Option to add Maxwellian external source terms
Source terms with a Maxwellian velocity profile (whose temperature is an
input option) with various options for the spatial profiles (currently
constant or Gaussian - with a tunable width - in each of the r- or
z-directions).

Includes an option for a PI controller for density.  Controller sets
amplitude of external source term to push the density (either full
profile or the midpoint value) to a fixed value.