sqrt-error-handling: Catch errors from vth=sqrt(2*ppar/dens) calculation nicely
When running in the REPL, it is annoying that MPI.Abort() has to be
called when a simulation crashes. This commit catches the most common
crash (from the sqrt() in the calculation of vth), and does an
MPI.Allreduce() so that every process can throw an error. This is
unnecessary communication for large-scale runs, so not sure about
merging it into the master branch...