Makie: Only ever make runtime plots on global_rank[]==0
Otherwise every block tries to make plots. Then which plot ends up in
the latest_plots.png file is random, and there may be errors when
different processes try to write the same file.

A better solution might be to gather data to global_rank[]==0, or maybe
to write a separate plot from each block.