CVODE: Split rk_update!() into RK-specific part, and generic post-timestep part
After any kind of timestep we need to update boundary conditions, moment
constraints, derived moments. Move these parts into a new function
post_timestep!() so that they can be called independently of the
Runge-Kutta time-stepper. The only parts left in rk_update!() are the
actual Runge-Kutta update of the variables.