CVODE: Function to compute time derivative given a current state
This may be useful for MMS tests of a single time-point, or for calling
black-box time-stepper libraries. It is currently slightly inefficient
(which could be improved in future).

The function copy-pastes a bunch of code for boundary conditions, which
is not ideal but not sure there is a way around this. Even if we added
functions to apply boundary conditions to df/dt, they would generally
have to be different for the ones used to apply boundary conditions to f
(since the 'boundary condition' on df/dt just sets it to 0.0 anywhere
that f is updated by a boundary condition rather than the time solver).