restart_interp: If restarting in a different run directory, do not move existing files
Sometimes it is useful to restart one run from a separate run (whose
output is in a different directory). In this case there is no need to
move the output files that are being read from, as they would not be
overwritten by the new run. So do not move them, to keep the directory
of the run being restarted from in its original state.