external-sources: Fix amplitude of integrated source terms
Had missed a factor of sqrt(vth). Most convenient way to fix is to
redefine source_amplitude to be the amplitude of a Maxwellian (which
integrates to 1) rather than of exp(-vpa^2/source_T), so the sqrt(vth)
factor is not needed in the integrated source terms, but has to be
divided out when the vpa-dependent source is added to the kinetic
equation (which is convenient as this happens in only one pair of
functions in external_sources).