Automatically include all dependencies when precompiling
In precompile.jl, precompile-no-run.jl and
precompile-makie-post-processing.jl, we can just use the default of
create_sysimage(), which is to include all packages associated with the
project.

For precompile_dependencies.jl where we explicitly want to exclude
moment_kinetics from the system image, we do need to pass a list of
packages, but we can read them from the Project.toml instead of
maintaing a separate, hard-coded list in the script.