Initialize neutral_vz_advect[isn].speed to zero
In full-f or evolving-density modes, neutral_vz_advect[isn].speed is not
needed for the 'vz advection' so update_speed_neutral_vz!() does
nothing. However, it may be used for the vz boundary condition, so
should be initialised. Setting it to 0.0 makes sense, because it makes
the 'incoming' boundary conditions ('incoming_zero' and, when there is
no diffusion 'zero') do nothing.