qz-neutral-bugfix: Fix typo so peculiar velocity grid is used to calculate qz_neutral
Previously, although the peculiar velocity was calculated in vz.scratch,
vz.grid was used as the integration variable, so the z-heat-flux was
wrong. This only affected cases without evolving-upar, so only affected
cases where qz_neutral is a diagnostic variable, not one actually used
to evolve the system. [The exception would be evolving-n-and-ppar, but
that case is not fully supported/implemented anyway.]