Fix use of itime_min/max/skip when using `it` to load subset of time points
Previously `it` was treated as the global time index in
postproc_load_data(), but this is surprising as it means that plots that
load data themselves use `it0` as a global index, whereas plots made
with data that was loaded first (using `itime_min`, `itime_max` and
`itime_skip`) use `it0` as a local index. Using `itime_min`, `itime_max`
and `itime_skip` always when handling `it` in `postproc_load_data()`
makes the two cases consistent.