makie-fixes: Make `rasterize` user-settable in f_unnorm plots
`rasteize` is needed as the plots produced by Makie's `mesh!()` function
result in huge file sizes when vectorized. However, the default setting
with just `rasterize=true` leads to fairly coarse plots. To get higher
resolution, a scale factor is passed to `rasterize=` instead of a
`Bool`. This commit adds a `rasterize` argument to
`plot_f_unnorm_vs_vpa_z()` so that the user can set this scale factor if
desired.