decaying-outgoing-bc: "zero_in_decaying_out" boundary condition for velocity coordinates
This boundary condition sets the incoming boundary to 0, like "zero",
but at the outgoing boundary forces the boundary point to be less than
or equal to the next point. This is an attempt to force the gradient to
be decreasing at the boundary - not guaranteed to work exactly as a
'better' implementation would require something like calculating the
derivative in the element (using all points) and working out the value
for the boundary point that would make the derivative zero, then forcing
the boundary point to be less than or equal to that value.