Simplify restart functionality
Move actual implementation of restart functionality inside
setup_moment_kinetics!() - this also fixes a bug where `iblock_index[]`
was not initialised when `get_backup_filename()` needed to use it.

This allows the `restart_moment_kinetics()` function to be removed. It
is replaced by a couple of arguments to run_moment_kinetics.

The `restart_moment_kinetics.jl` script is removed. Now either pass the
`--restart` flag to `run_moment_kinetics()` to restart from the latest
file in the run directory, or pass the output file to restart from as
the second positional argument.