`discretization_info` types for length-1 dimensions
Length-1 dimensions have to be treated a bit specially. Derivatives in
those directions should be zero. 'Interpolation' to a grid with more
than one point: for spatial dimensions, assume the variable is constant;
for velocity dimensions, use a Maxwellian whose peak value is the value
on the length-1 grid, and whose width is 1 (in units of the reference
speed) - this means that the integral of the variable over velocity
space, after accounting for factors of pi^0.5 or pi^1.5 in the
normalization of distribution functions, is the same (up to
discretization error) after 'interpolating' as it was before.