Split 'Post processing' docs and `makie_post_processing` API
Putting these in separate files means both show up in the contents list
in the left sidebar. This is better because it means that all modules
show up directly (previously `makie_post_processing` was a subsection of
the 'Post processing' page, which would make it harder to find if you
expected a direct link to the module API documentation). Now the 'API'
section of the 'Post processing' page is just a link to the page with
the `makie_post_processing` API docs.