Rename docs files to fix sorting of contents list
The sorting of the pages in the left sidebar of the docs is done by
filename, not page title (although the page titles are displayed). To
make this less confusing, rename some files to correspond more closely
to the page titles, so that the sidebar looks like it's sorted by page
title.

Also prefix all API docs files with 'zz_' so that they get displayed
after the main docs pages.