eftc2023-jto: Use sqrt-spaced z-grid for EFTC examples
...otherwise numerical issues in the first/last elements in z prevent
the 'split3' simulations from converging to a state with residuals below
1e-3 (but all input files are changed, not just the 'split3' ones, as
the 'split3' simulations need to restart from a case with a sqrt-spaced
grid to avoid numerical instability). Adds a note in the README.md that
this grid was not used for the plots on the EFTC poster, these were just
made accepting slightly higher residuals for the 'split3' case (the
difference was in any case not visible on the plots).