Just warn in get_run_info() if default input file does not exist
It should not be necessary to have a post-processing input .toml file to
call `get_run_info()` interactively. Previously a check for a missing
input file in `setup_makie_post_processing()` would raise an error if
the specified input file (which might be the default input file) was not
present. This commit adds an argument to replace that error with a
warning, and this argument is passed in `get_run_info()` if no input
file was explicitly passed using the `setup_input_file` argument (if a
file was passed, presumably the user expects it to exist, so an error is
helpful, e.g. if there was a typo in the filename).