Exclude version 1.14 of HDF5_jll package
Version 1.14 of the HDF5_jll package causes a bug (see #148), so set the
`[compat]` section of Project.toml to exclude this version (hoping the
bug will be fixed by version 1.15 at the latest, some future patch
release like 1.14.3 might also be OK though).

HDF5_jll is not used directly in moment_kinetics, and so is only
included in the Project.toml in order to apply this version restriction.
It should probably be removed once the latest version has fixed the bug.