Get temporary output directory at run time in all tests
Previously, temporary directory was created at pre-compile time and its
name saved as a module variable, so would be the same for multiple runs
of the test. That would make deleting the output directory to save disk
space more complicated. Better to create the temporary directory at run
time, so it is different for each test.