When possible, throw error rather than calling MPI.Abort()
If an error occurs on all processes at once, then we can just throw the
error - there is no need to call MPI.Abort(). This is nice for
interactive usage, where for example now a typo in some option in an
input file will not kill Julia.

This commit adds a function `timeout_mpi_barrier()` which can be used to
test whether all processes call the function within some timeout
interval (by default 5 seconds), so that if some processes errored and
some did not then we can still call MPI.Abort() to stop the program from
hanging.