Fix @debug_track_initialized for running in parallel
The array that tracks whether each element is initialized must also be a
shared-memory array, otherwise (which was happening before this commit)
an error would be raised if an element was initialized on one process
but read on another. This also means we have to use an array of `mk_int`
for `is_initialized`, rather than an array of `Bool`, because MPI does
not have a native Bool type, so we cannot create a shared array of
`Bool`.