Reduce number of spatial grid points in FP collision operator debug test
The Fokker-Planck collisions do not do any spatially-dependent
operations, so having just nelement=1, ngrid=2 in the r- and z-diretions
should be enough to check for shared-memory errors. Reducing the number
of spatial grid points from 5*5=25 to 2*2=4 massively speeds up the
debug test, making it more practical to run (especially on the CI
servers).